fixed as suggested
