+++ packages/m2m-gateway/src/api/agreementApiConverter.ts
@@ -55,3 +55,15 @@ export function toGetPurposesApiQueryParamsForAgreement(
     name: undefined,
   };
 }
+
+export function toM2MGatewayApiDocument(
+  document: agreementApi.Document
+): m2mGatewayApi.Document {
+  return {
+    id: document.id,
+    name: document.name,
+    prettyName: document.prettyName,
+    createdAt: document.createdAt,
+    contentType: document.contentType,
+  };
+}
