+++ packages/m2m-gateway/src/api/eserviceApiConverter.ts
@@ -107,3 +107,15 @@ export function toM2MGatewayApiEServiceRiskAnalysis(
     ),
   };
 }
+
+export function toM2MGatewayApiDocument(
+  document: catalogApi.EServiceDoc
+): m2mGatewayApi.Document {
+  return {
+    id: document.id,
+    name: document.name,
+    prettyName: document.prettyName,
+    createdAt: document.uploadDate,
+    contentType: document.contentType,
+  };
+}
